Separate out alpha endpoints
