+++ graphdatascience/graph/graph_endpoints.py
@@ -1,4 +1,5 @@
 from ..caller_base import CallerBase
+from .graph_alpha_proc_runner import GraphAlphaProcRunner
 from .graph_proc_runner import GraphProcRunner
 
 
@@ -6,3 +7,9 @@ class GraphEndpoints(CallerBase):
     @property
     def graph(self) -> GraphProcRunner:
         return GraphProcRunner(self._query_runner, f"{self._namespace}.graph", self._server_version)
+
+
+class GraphAlphaEndpoints(CallerBase):
+    @property
+    def graph(self) -> GraphAlphaProcRunner:
+        return GraphAlphaProcRunner(self._query_runner, f"{self._namespace}.graph", self._server_version)
