+++ graphdatascience/model/model_endpoints.py
@@ -1,4 +1,5 @@
 from ..caller_base import CallerBase
+from .model_alpha_proc_runner import ModelAlphaProcRunner
 from .model_proc_runner import ModelProcRunner
 
 
@@ -6,3 +7,9 @@ class ModelEndpoints(CallerBase):
     @property
     def model(self) -> ModelProcRunner:
         return ModelProcRunner(self._query_runner, f"{self._namespace}.model", self._server_version)
+
+
+class ModelAlphaEndpoints(CallerBase):
+    @property
+    def model(self) -> ModelAlphaProcRunner:
+        return ModelAlphaProcRunner(self._query_runner, f"{self._namespace}.model", self._server_version)
