+++ graphdatascience/pipeline/pipeline_endpoints.py
@@ -2,6 +2,7 @@ from ..caller_base import CallerBase
 from .lp_training_pipeline import LPTrainingPipeline
 from .nc_training_pipeline import NCTrainingPipeline
 from .nr_training_pipeline import NRTrainingPipeline
+from .pipeline_alpha_proc_runner import PipelineAlphaProcRunner
 from .pipeline_proc_runner import PipelineProcRunner
 
 
@@ -21,6 +22,12 @@ class PipelineEndpoints(CallerBase):
         return p
 
     def nr_pipe(self, name: str) -> NRTrainingPipeline:
-        runner = PipelineProcRunner(self._query_runner, f"{self._namespace}.alpha.pipeline", self._server_version)
+        runner = PipelineAlphaProcRunner(self._query_runner, f"{self._namespace}.alpha.pipeline", self._server_version)
         p, _ = runner.nodeRegression.create(name)
         return p
+
+
+class PipelineAlphaEndpoints(CallerBase):
+    @property
+    def pipeline(self) -> PipelineAlphaProcRunner:
+        return PipelineAlphaProcRunner(self._query_runner, f"{self._namespace}.pipeline", self._server_version)
